+++ game/shared/teammgr.cpp
@@ -377,13 +377,13 @@ uint8 CTeamMgr::AddTeam(const char *szName, uint32 nModel)
 	if( !pTeam )
 		return INVALID_TEAM;
 
-	pTeam->Init(m_teams.size(),szName,nModel);
+	pTeam->Init(static_cast<uint8>(m_teams.size()),szName,nModel);
 
 	m_teams.push_back(pTeam);
 
 g_pLTServer->CPrint( "Added team '%s' with model (%i)", szName, nModel );
 
-	return m_teams.size();
+	return static_cast<uint8>(m_teams.size());
 
 }
